requirements.txt
bcrypt==3.2.0
mysql-connector-python==8.0.30
requests==2.28.1